break apart GitRemoteConnection into parse/write functions
