+++ src/internal-apis.js
@@ -6,7 +6,6 @@ export * from './managers/GitConfigManager'
 export * from './managers/GitIgnoreManager'
 export * from './managers/GitIndexManager'
 export * from './managers/GitRefManager'
-export * from './managers/GitRemoteConnection'
 export * from './managers/GitRemoteHTTP'
 export * from './managers/GitRemoteManager'
 export * from './managers/GitShallowManager'
@@ -45,3 +44,9 @@ export * from './utils/shasum'
 export * from './utils/sleep'
 export * from './utils/symbols'
 export * from './utils/t'
+
+export * from './wire/parseRefsAdResponse'
+export * from './wire/parseUploadPackResponse'
+export * from './wire/parseReceivePackResponse'
+export * from './wire/writeUploadPackRequest'
+export * from './wire/writeReceivePackRequest'
